+++ client-mr/core/src/test/java/org/apache/hadoop/mapreduce/task/reduce/EventFetcherTest.java
@@ -58,7 +58,7 @@ public class EventFetcherTest {
     RssEventFetcher ef = new RssEventFetcher(1, tid, umbilical, jobConf, MAX_EVENTS_TO_FETCH);
     Roaring64NavigableMap expected = Roaring64NavigableMap.bitmapOf();
     for (int mapIndex = 0; mapIndex < mapTaskNum; mapIndex++) {
-      int rssTaskId =
+      long rssTaskId =
           RssMRUtils.createRssTaskAttemptId(
               new TaskAttemptID("12345", 1, TaskType.MAP, mapIndex, 0), 1, 4);
       expected.addLong(rssTaskId);
@@ -88,7 +88,7 @@ public class EventFetcherTest {
     RssEventFetcher ef = new RssEventFetcher(1, tid, umbilical, jobConf, MAX_EVENTS_TO_FETCH);
     Roaring64NavigableMap expected = Roaring64NavigableMap.bitmapOf();
     for (int mapIndex = 0; mapIndex < mapTaskNum; mapIndex++) {
-      int rssTaskId =
+      long rssTaskId =
           RssMRUtils.createRssTaskAttemptId(
               new TaskAttemptID("12345", 1, TaskType.MAP, mapIndex, 0), 1, 4);
       expected.addLong(rssTaskId);
@@ -120,7 +120,7 @@ public class EventFetcherTest {
 
     Roaring64NavigableMap expected = Roaring64NavigableMap.bitmapOf();
     for (int mapIndex = 0; mapIndex < mapTaskNum; mapIndex++) {
-      int rssTaskId =
+      long rssTaskId =
           RssMRUtils.createRssTaskAttemptId(
               new TaskAttemptID("12345", 1, TaskType.MAP, mapIndex, 0), 1, 4);
       expected.addLong(rssTaskId);
@@ -145,7 +145,7 @@ public class EventFetcherTest {
     RssEventFetcher ef = new RssEventFetcher(1, tid, umbilical, jobConf, MAX_EVENTS_TO_FETCH);
     Roaring64NavigableMap expected = Roaring64NavigableMap.bitmapOf();
     for (int mapIndex = 0; mapIndex < mapTaskNum; mapIndex++) {
-      int rssTaskId =
+      long rssTaskId =
           RssMRUtils.createRssTaskAttemptId(
               new TaskAttemptID("12345", 1, TaskType.MAP, mapIndex, 0), 1, 4);
       expected.addLong(rssTaskId);
@@ -171,7 +171,7 @@ public class EventFetcherTest {
     RssEventFetcher ef = new RssEventFetcher(1, tid, umbilical, jobConf, MAX_EVENTS_TO_FETCH);
     Roaring64NavigableMap expected = Roaring64NavigableMap.bitmapOf();
     for (int mapIndex = 0; mapIndex < mapTaskNum; mapIndex++) {
-      int rssTaskId =
+      long rssTaskId =
           RssMRUtils.createRssTaskAttemptId(
               new TaskAttemptID("12345", 1, TaskType.MAP, mapIndex, 0), 1, 4);
       expected.addLong(rssTaskId);
@@ -204,13 +204,13 @@ public class EventFetcherTest {
     Roaring64NavigableMap expected = Roaring64NavigableMap.bitmapOf();
     for (int mapIndex = 0; mapIndex < mapTaskNum; mapIndex++) {
       if (!tipFailed.contains(mapIndex) && !obsoleted.contains(mapIndex)) {
-        int rssTaskId =
+        long rssTaskId =
             RssMRUtils.createRssTaskAttemptId(
                 new TaskAttemptID("12345", 1, TaskType.MAP, mapIndex, 0), 1, 4);
         expected.addLong(rssTaskId);
       }
       if (obsoleted.contains(mapIndex)) {
-        int rssTaskId =
+        long rssTaskId =
             RssMRUtils.createRssTaskAttemptId(
                 new TaskAttemptID("12345", 1, TaskType.MAP, mapIndex, 1), 1, 4);
         expected.addLong(rssTaskId);
