+++ client-spark/spark2/src/main/java/org/apache/spark/shuffle/RssShuffleManager.java
@@ -468,7 +468,7 @@ public class RssShuffleManager extends RssShuffleManagerBase {
   }
 
   @Override
-  public int getTaskAttemptIdForBlockId(int mapIndex, int attemptNo) {
+  public long getTaskAttemptIdForBlockId(int mapIndex, int attemptNo) {
     return getTaskAttemptIdForBlockId(
         mapIndex, attemptNo, maxFailures, speculation, blockIdLayout.taskAttemptIdBits);
   }
