+++ client-spark/spark3/src/main/java/org/apache/spark/shuffle/RssShuffleManager.java
@@ -546,7 +546,7 @@ public class RssShuffleManager extends RssShuffleManagerBase {
   }
 
   @Override
-  public int getTaskAttemptIdForBlockId(int mapIndex, int attemptNo) {
+  public long getTaskAttemptIdForBlockId(int mapIndex, int attemptNo) {
     return getTaskAttemptIdForBlockId(
         mapIndex, attemptNo, maxFailures, speculation, blockIdLayout.taskAttemptIdBits);
   }
